Add  error handling and validation

+++ src/controllers/movieController.js
@@ -3,6 +3,7 @@ import { Router } from "express";
 import movieService from "../services/movieService.js";
 import castService from "../services/castService.js";
 import { isAuth } from "../middlewares/authMiddleware.js";
+import { getErrorMessage } from "../utils/errorUtils.js";
 
 const router = Router();
 
@@ -15,7 +16,14 @@ router.post('/create', isAuth, async (req, res) => {
     const movieData = req.body;
     const ownerId = req.user?._id;
 
-    await movieService.create(movieData, ownerId);
+    try {
+        await movieService.create(movieData, ownerId);
+    } catch (err) {
+        // Challenge: Show multi errors
+        const errorMessage = getErrorMessage(err);
+
+        return res.render('movies/create', { error: errorMessage, movie: movieData });
+    }
 
     res.redirect('/');
 });
@@ -55,6 +63,14 @@ router.post('/:movieId/attach', isAuth, async (req, res) => {
 
 router.get('/:movieId/delete', isAuth, async (req, res) => {
     const movieId = req.params.movieId;
+
+    // Check if owner
+    const movie = await movieService.getOne(movieId).lean();
+    if (movie.owner?.toString() !== req.user._id) {
+        // return res.render('movies/details', { movie, isOwner: false, error: 'You cannot delete this movie!' });
+        res.setError('You cannot delete this movie!');
+        return res.redirect('/404');
+    }
 
     await movieService.remove(movieId);
